Created sample motion
